# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Manus AI API
  description: |
    A comprehensive, versatile AI system built with NestJS that provides specialized AI agents capable of performing various tasks through real AI integration (OpenAI + LangChain), background processing, and system automation.

    ## Key Features
    - **9 Specialized AI Agents** with real AI integration (no mock data)
    - **Background Task Processing** with BullMQ and Redis queues
    - **Session Persistence** for offline capability and long-running tasks
    - **Real-time AI Interactions** with streaming responses via LangChain
    - **System Tool Integration** (file operations, browser automation, shell commands)

    ## Technologies Used
    - **OpenAI + LangChain**: Real AI integration with proper streaming and tool binding
    - **NestJS**: Progressive Node.js framework
    - **BullMQ**: Advanced Redis-based queue system
    - **MongoDB**: NoSQL database with Prisma ORM
    - **TypeScript**: Strongly typed development

  version: 1.0.0
  contact:
    name: Manus AI Team
    email: support@manus.im
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.manus.im/api/v1
    description: Production server
paths:
...
  /edit:
    post:
      summary: General file editing
      description: Perform general file editing operations using AI
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRequest'
      responses:
        '202':
          description: Edit task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/create:
    post:
      summary: Create new file
      description: Create a new file with AI-generated content
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileCreateRequest'
      responses:
        '202':
          description: File creation task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/update:
    post:
      summary: Update existing file
      description: Update an existing file using AI assistance
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileUpdateRequest'
      responses:
        '202':
          description: File update task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/delete:
    post:
      summary: Delete file
      description: Delete a file from the system
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileDeleteRequest'
      responses:
        '202':
          description: File deletion task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/refactor:
    post:
      summary: Refactor code
      description: Refactor existing code using AI assistance
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeRefactorRequest'
      responses:
        '202':
          description: Code refactoring task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/format:
    post:
      summary: Format code
      description: Format code using AI-powered formatting
      tags:
        - Edit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeFormatRequest'
      responses:
        '202':
          description: Code formatting task queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /edit/history/{sessionId}:
    get:
      summary: Get edit history
      description: Get the edit history for a session
      tags:
        - Edit
        - History
      parameters:
        - name: sessionId
          in: path
          required: true
          description: Session ID
          schema:
            type: string
      responses:
        '200':
          description: Edit history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EditHistoryItem'
  /edit/task/{taskId}/retry:
    post:
      summary: Retry failed edit task
      description: Retry a failed edit task
      tags:
        - Edit
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task retried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/task/{taskId}/cancel:
    post:
      summary: Cancel edit task
      description: Cancel a running edit task
      tags:
        - Edit
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
  /edit/task/{taskId}:
    delete:
      summary: Delete edit task
      description: Delete a completed or failed edit task
      tags:
        - Edit
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  schemas:
...
    TaskResponse:
      type: object
      properties:
        taskId:
          type: string
          description: Unique task identifier
        status:
          type: string
          enum: [queued, processing, completed, failed]
        estimatedTime:
          type: string
          description: Estimated completion time
        sessionId:
          type: string
...
    EditRequest:
      type: object
      required:
        - instruction
        - sessionId
      properties:
        instruction:
          type: string
          description: Edit instruction
        target:
          type: string
          description: Target file or content
        sessionId:
          type: string
    FileCreateRequest:
      type: object
      required:
        - instruction
        - sessionId
      properties:
        instruction:
          type: string
          description: File creation instruction
        filePath:
          type: string
          description: Desired file path
        sessionId:
          type: string
    FileUpdateRequest:
      type: object
      required:
        - instruction
        - target
        - sessionId
      properties:
        instruction:
          type: string
          description: Update instruction
        target:
          type: string
          description: Target file path
        sessionId:
          type: string
    FileDeleteRequest:
      type: object
      required:
        - target
        - sessionId
      properties:
        target:
          type: string
          description: File to delete
        sessionId:
          type: string
    CodeRefactorRequest:
      type: object
      required:
        - target
        - instruction
        - sessionId
      properties:
        target:
          type: string
          description: Target file or code to refactor
        operation:
          type: string
          default: refactor
        instruction:
          type: string
          description: Refactoring instruction
        sessionId:
          type: string
    CodeFormatRequest:
      type: object
      required:
        - code
        - sessionId
      properties:
        code:
          type: string
          description: Code to format
        language:
          type: string
          description: Programming language
        sessionId:
          type: string
...
    EditHistoryItem:
      type: object
      properties:
        id:
          type: string
        operation:
          type: string
          enum: [create, update, delete, refactor, format]
        target:
          type: string
        instruction:
          type: string
        result:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: string
          enum: [completed, failed]
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...